fix: address V.2 QA-1 findings
