+++ crates/atm-daemon/src/daemon_observability.rs
@@ -10,11 +10,11 @@ use std::{fs, fs::OpenOptions};
 use atm_core::error::AtmError;
 use atm_core::error_codes::AtmErrorCode;
 use atm_core::home;
-use atm_core::schema::AtmMessageId;
 use atm_core::observability::{
     AtmLogQuery, AtmLogSnapshot, AtmObservabilityHealth, AtmObservabilityHealthState, CommandEvent,
     LogTailSession, ObservabilityPort, RetainedSinkFaultMode,
 };
+use atm_core::schema::AtmMessageId;
 use sc_observability::{
     LogSink, Logger, LoggerConfig, RetentionPolicy, RotationPolicy, SinkRegistration,
 };
@@ -333,10 +333,12 @@ impl RetainedJsonlFileSink {
     fn schedule_prune_old_files(&self) {
         let now = SystemTime::now();
         {
-            let mut last_request_at = self
-                .last_prune_request_at
-                .lock()
-                .expect("retained sink prune request lock poisoned");
+            let Ok(mut last_request_at) = self.last_prune_request_at.lock() else {
+                tracing::warn!(
+                    "retained sink prune request lock poisoned; skipping one prune scheduling attempt"
+                );
+                return;
+            };
             if let Some(last_request_at) = *last_request_at
                 && now
                     .duration_since(last_request_at)
@@ -384,9 +386,12 @@ impl RetainedJsonlFileSink {
         )
         .cause(message)
         .source(Box::new(error));
-        let mut health = self.health.lock().expect("file sink health poisoned");
-        health.state = SinkHealthState::DegradedDropping;
-        health.last_error = Some(DiagnosticSummary::from(diagnostic.diagnostic()));
+        if let Ok(mut health) = self.health.lock() {
+            health.state = SinkHealthState::DegradedDropping;
+            health.last_error = Some(DiagnosticSummary::from(diagnostic.diagnostic()));
+        } else {
+            tracing::warn!("file sink health lock poisoned while recording sink failure");
+        }
         LogSinkError(Box::new(diagnostic))
     }
 }
@@ -439,20 +444,24 @@ impl LogSink for RetainedJsonlFileSink {
         file.write_all(&line)
             .and_then(|()| file.flush())
             .map_err(|error| self.mark_failure(error))?;
-        *self
-            .last_written_file
-            .lock()
-            .expect("retained sink file handle poisoned") = Some(file);
-        let mut health = self.health.lock().expect("file sink health poisoned");
+        *self.last_written_file.lock().map_err(|_| {
+            self.mark_failure(std::io::Error::other(
+                "retained sink file handle lock poisoned",
+            ))
+        })? = Some(file);
+        let mut health = self.health.lock().map_err(|_| {
+            self.mark_failure(std::io::Error::other("file sink health lock poisoned"))
+        })?;
         health.state = SinkHealthState::Healthy;
         Ok(())
     }
 
     fn flush(&self) -> Result<(), LogSinkError> {
-        let mut last_written = self
-            .last_written_file
-            .lock()
-            .expect("retained sink file handle poisoned");
+        let mut last_written = self.last_written_file.lock().map_err(|_| {
+            self.mark_failure(std::io::Error::other(
+                "retained sink file handle lock poisoned",
+            ))
+        })?;
         if let Some(file) = last_written.as_mut() {
             file.flush().map_err(|error| self.mark_failure(error))?;
         }
@@ -460,10 +469,17 @@ impl LogSink for RetainedJsonlFileSink {
     }
 
     fn health(&self) -> SinkHealth {
-        self.health
-            .lock()
-            .expect("file sink health poisoned")
-            .clone()
+        match self.health.lock() {
+            Ok(health) => health.clone(),
+            Err(_) => {
+                tracing::warn!("file sink health lock poisoned; reporting unavailable sink health");
+                SinkHealth {
+                    name: SinkName::new("jsonl_file_sink").expect("jsonl sink constant is valid"),
+                    state: SinkHealthState::Unavailable,
+                    last_error: None,
+                }
+            }
+        }
     }
 }
 
@@ -679,10 +695,16 @@ fn map_daemon_event(
     );
     match &event.team {
         TeamScope::Team(team) => {
-            fields.insert("team".to_string(), serde_json::Value::String(team.to_string()));
+            fields.insert(
+                "team".to_string(),
+                serde_json::Value::String(team.to_string()),
+            );
         }
         TeamScope::None => {
-            fields.insert("team_scope".to_string(), serde_json::Value::String("none".to_string()));
+            fields.insert(
+                "team_scope".to_string(),
+                serde_json::Value::String("none".to_string()),
+            );
         }
     }
     if let Some(agent) = event.agent.as_ref() {
@@ -843,7 +865,7 @@ mod tests {
     use tempfile::TempDir;
 
     use super::DaemonObservability;
-    use atm_daemon::{DaemonEvent, DaemonSubsystem, TeamScope};
+    use atm_daemon::observability_test_event;
 
     #[test]
     #[serial]
@@ -859,32 +881,18 @@ mod tests {
 
         let observability = DaemonObservability::bootstrap().expect("bootstrap");
         observability
-            .emit_daemon_event(DaemonEvent {
-                subsystem: DaemonSubsystem::Composition,
-                action: "start_requested",
-                outcome: "ok",
-                team: TeamScope::None,
-                agent: None,
-                sender: None,
-                recipient: None,
-                message_id: None,
-                task_id: None,
-                detail: "daemon start requested".into(),
-            })
+            .emit_daemon_event(observability_test_event(
+                "start_requested",
+                "ok",
+                "daemon start requested",
+            ))
             .expect("emit start");
         observability
-            .emit_daemon_event(DaemonEvent {
-                subsystem: DaemonSubsystem::Composition,
-                action: "shutdown_completed",
-                outcome: "ok",
-                team: TeamScope::None,
-                agent: None,
-                sender: None,
-                recipient: None,
-                message_id: None,
-                task_id: None,
-                detail: "daemon shutdown completed".into(),
-            })
+            .emit_daemon_event(observability_test_event(
+                "shutdown_completed",
+                "ok",
+                "daemon shutdown completed",
+            ))
             .expect("emit shutdown");
         observability
             .best_effort_flush_blocking()
@@ -988,18 +996,11 @@ mod tests {
             )
             .expect("bootstrap");
         observability
-            .emit_daemon_event(DaemonEvent {
-                subsystem: DaemonSubsystem::Composition,
-                action: "start_requested",
-                outcome: "ok",
-                team: TeamScope::None,
-                agent: None,
-                sender: None,
-                recipient: None,
-                message_id: None,
-                task_id: None,
-                detail: "daemon start requested".into(),
-            })
+            .emit_daemon_event(observability_test_event(
+                "start_requested",
+                "ok",
+                "daemon start requested",
+            ))
             .expect("emit");
 
         let active_log_path = log_dir.join("atm.log.jsonl");
