+++ crates/atm-daemon/src/lib.rs
@@ -40,6 +40,8 @@ use std::time::Duration;
 
 use atm_core::error::{AtmError, AtmErrorCode};
 use atm_rusqlite::SqliteBoundaryAssembly;
+#[cfg(test)]
+pub(crate) use daemon_runtime_observability::observability_test_event;
 pub use daemon_runtime_observability::{
     DaemonEvent, DaemonRuntimeObservability, DaemonSubsystem, TeamScope,
 };
